fix(routing): stop dragging the wake subsystem into resolveTarget — unblocks CI
CI on v26.6.5-alpha.1323 failed with "SyntaxError: Export named 'hostExec'
not found in module src/sdk/index.ts" in comm-send-focused-gaps.test.ts,
blocking the maw-team-up release.

Root cause (traced, NOT a circular import / export-ordering bug):
  comm-send → routing (detectWindowMismatch) → imports resolveFleetSession
  from the `wake` BARREL → which eagerly re-exports wake-session +
  wake-resolve-impl, both of which `import { hostExec, tmux } from "../../sdk"`.
A test that partial-mocks the sdk barrel (omitting hostExec) then fails to
LINK that named import. routing is hot (comm-send, /api/send), so it dragged
the entire transport-coupled wake subsystem into every consumer's graph.

Fix: resolveFleetSession is a pure fleet-config lookup (loadFleet +
resolveFleetWindowSessionTarget). Move it to the light `fleet-load` module
(which only uses `import * as sdk` — a namespace import that is link-safe
under partial mocks) and point routing there. wake-resolve-impl re-exports it
so the wake barrel and all other importers are unchanged. This removes
wake-session/wake-cmd/wake-resolve-impl from routing's graph entirely,
shrinking the mock surface for routing/comm-send tests from 3 barrels to 0.

Also completes the test's intended oracle-manifest mock (loadManifestCached),
which was masked behind the hostExec failure.

Verified: comm-send-focused-gaps 5/5; routing/comm-send/wake/fleet/sessions-api
suites green individually; tsc clean; bun run build OK (688 modules, no cycle).
